split parser files up into 2 files: old csv and new csv parser (and tests)
