+++ backend/src/csv_parsing/csv_file_parser.ts
@@ -17,23 +17,6 @@ import {
 } from "@repo/common";
 
 /**
- * Type used to structure production import
- */
-type ParsedProductionImport = {
-  productions: (CreateProductionDto & { legacy_id: string })[];
-  tags: string[];
-  productionTagLinks: { legacyId: string; tagName: string }[];
-};
-
-/**
- * Internal helper used when parsing events.
- */
-type ParsedEventRow = {
-  event: CreateEventDto;
-  location: string;
-};
-
-/**
  * Internal helper used when parsing localized events.
  */
 type ParsedLocalizedEventRow = {
@@ -75,17 +58,6 @@ type ParsedTagRow = {
 };
 
 export class CSVFileParser {
-  private static toOldLocalizedString(value: string): {
-    en: string;
-    nl: string;
-  } {
-    const normalized = (value || "").trim();
-    return {
-      en: normalized,
-      nl: normalized,
-    };
-  }
-
   private static toLocalizedString(
     value_nl: string,
     value_en: string,
@@ -167,75 +139,6 @@ export class CSVFileParser {
         cleanupAndReject(new Error(`Error reading CSV file: ${error.message}`));
       });
     });
-  }
-
-  static transformOldEventRow(
-    row: Record<string, string>,
-  ): CreateEventDto & { location: string } {
-    let endTime: string | null = null;
-
-    // Check for invalid date formats and handle them accordingly
-    const invalidDates = ["0000-00-00 00:00:00", "1970-01-01 00:00:00", ""];
-    const starttimeDate = new Date(row.Starttime);
-    if (isNaN(starttimeDate.getTime())) {
-      throw new Error(`Invalid starttime: ${row.Starttime}`);
-    }
-    if (row.Endtime && !invalidDates.includes(row.Endtime)) {
-      const endTimeDate = new Date(row.Endtime);
-      if (!isNaN(endTimeDate.getTime()) && endTimeDate > starttimeDate)
-        endTime = endTimeDate.toISOString();
-    }
-
-    const productionId = Number(row.Production);
-    if (isNaN(productionId)) {
-      throw new Error(`Invalid production id: ${row.Production}`);
-    }
-
-    const location = (row.Hall || "").trim();
-
-    return {
-      starttime: starttimeDate.toISOString(),
-      endtime: endTime,
-      production_id: productionId,
-      location: location,
-      doors_at: null, // TODO
-      intermission_at: null, // TODO
-    };
-  }
-
-  static transformOldProductionRow(
-    row: Record<string, string>,
-  ): CreateProductionDto & { tags: string[]; legacy_id: string } {
-    // validate and convert numeric fields manually to provide clearer errors
-    const id = Number(row.ID);
-    if (isNaN(id)) {
-      throw new Error(`Invalid production id: ${row.ID}`);
-    }
-
-    // adds possibility to add tags as a comma separated list in the Genre column, trims whitespace and converts to lowercase for consistency
-    const tags = (row.Genre || "")
-      .split(",")
-      .map((t) => t.trim().toLowerCase())
-      .filter(Boolean);
-
-    const tagLine = row.Tagline || row.Ondertitel || null;
-
-    const legacy_id = `csv-${id}`;
-
-    return {
-      titel: CSVFileParser.toOldLocalizedString(row.Titel),
-      description1: CSVFileParser.toOldLocalizedString(row.Description1),
-      description2: row.Description2
-        ? CSVFileParser.toOldLocalizedString(row.Description2)
-        : null,
-      artist: null, // TODO
-      tagline: tagLine ? CSVFileParser.toOldLocalizedString(tagLine) : null,
-      credits: null, // TODO
-      attendance_mode: null, // TODO
-      performer_type: null, // TODO
-      legacy_id: legacy_id,
-      tags: [...new Set(tags)], // remove duplicate tags
-    };
   }
 
   static transformEventRow(row: Record<string, string>): CreateEventDto & {
@@ -421,71 +324,6 @@ export class CSVFileParser {
     return {
       tag,
       productionIds,
-    };
-  }
-
-  /**
-   * Parse events from an old CSV file and return them along with a raw location
-   * string.
-   * @param filePath - Path to the CSV file containing events
-   * @return A promise that resolves to an array of objects containing the
-   *         event DTO and the value of the `Hall` column (may be empty).
-   */
-  static async parseOldEventsCSV(filePath: string): Promise<ParsedEventRow[]> {
-    const parsed = await this.parseCSVWithSchema<
-      CreateEventDto & { location: string }
-    >(
-      filePath,
-      CreateEventSchema.extend({ location: string() }),
-      this.transformOldEventRow,
-    );
-
-    return parsed.map((r) => {
-      const { location, ...eventData } = r;
-      return { event: eventData, location: location };
-    });
-  }
-
-  /**
-   * Parse productions from an old CSV file and return structured import data
-   * @param filePath - Path to the CSV file containing productions
-   * @return A promise that resolves to an object containing productions, unique tags, and production-tag links
-   */
-  static async parseOldProductionsCSV(
-    filePath: string,
-  ): Promise<ParsedProductionImport> {
-    const productions: (CreateProductionDto & { legacy_id: string })[] = [];
-    const tagsSet: Set<string> = new Set();
-    const productionTagLinks: { legacyId: string; tagName: string }[] = [];
-
-    const parsed = await this.parseCSVWithSchema<
-      CreateProductionDto & { tags: string[]; legacy_id: string }
-    >(
-      filePath,
-      CreateProductionSchema.extend({
-        tags: string().array(),
-        legacy_id: string(),
-      }),
-      this.transformOldProductionRow,
-    );
-
-    for (const production of parsed) {
-      const { tags, ...prodData } = production;
-      productions.push(prodData);
-
-      for (const tag of tags) {
-        tagsSet.add(tag);
-        productionTagLinks.push({
-          legacyId: prodData.legacy_id,
-          tagName: tag,
-        });
-      }
-    }
-
-    return {
-      productions: productions,
-      tags: Array.from(tagsSet),
-      productionTagLinks,
     };
   }
